requirements.txt
click==8.1.3
python-dotenv==1.0.0
web3==6.4.0
sqlalchemy==1.4.42
simplejson==3.17.6
pydantic==1.9.2
psycopg2-binary==2.9.6
pytest==7.3.1
pandas==2.0.0
